v1 phase 1b: state layout — id widening, SubVault v1 fields, bid tree live, LTV stamps
- sub_vault_id u8 -> u16 across every carrier (SubVault, ClaimedSeat,
  LoanFixed.lender_sub_vault_id, UserLoanRef.counterparty_sub_vault_id,
  GlobalVaultFixed.next_sub_vault_id/sub_vault_count, params, logs)
- SubVault gains kind / spread_bps / liquidation_ltv_bps /
  curator_fee_bps / open_orders_count / open_loans_count (zeroed;
  semantics land in later phases); counters folded into is_empty()
- MatchedLoan + LoanFixed gain origination_ltv_bps / liquidation_ltv_bps
  (LoanFixed reclaims the unused debt_index_at_last_accrual u128)
- MarketFixed _reserved_bids_* -> bids_root_index / bids_best_index
- ACCOUNT_LAYOUT_VERSION 1 -> 2; all struct sizes pinned by const asserts
- tests: 149 lib + 43 native + 140 SBF green

+++ programs/ydelta/src/program/instruction_builders/admin_transfer_instructions.rs
@@ -102,7 +102,7 @@ pub fn accept_global_vault_admin_instruction(mint: &Pubkey, pending_admin: &Pubk
 pub fn transfer_curator_instruction(
     mint: &Pubkey,
     current_curator: &Pubkey,
-    sub_vault_id: u8,
+    sub_vault_id: u16,
     new_curator: &Pubkey,
 ) -> Instruction {
     let (vault, _) = global_vault_pda(mint);
@@ -130,7 +130,7 @@ pub fn transfer_curator_instruction(
 pub fn accept_curator_instruction(
     mint: &Pubkey,
     pending_curator: &Pubkey,
-    sub_vault_id: u8,
+    sub_vault_id: u16,
 ) -> Instruction {
     let (vault, _) = global_vault_pda(mint);
     let mut data = YdeltaInstruction::AcceptCurator.to_vec();
